Bugfix for non-existent mesonh/ext directory

+++ check_commit_mesonh.sh
@@ -181,8 +181,10 @@ if [ $compilation -eq 1 ]; then
   find PHYEX -type f -exec touch {} \; #to be sure a recompilation occurs
 
   # Move manually ext/ files in src/MNH
-  mv -f PHYEX/ext/* MNH/
-  rmdir PHYEX/ext
+  if [ -d PHYEX/ext ]; then
+    mv -f PHYEX/ext/* MNH/
+    rmdir PHYEX/ext
+  ]
 
   cd $MNHPACK/$name/src/PHYEX/turb
   # Delete files of MNH-V5-5-0/src/MNH and MNH/src/LIB/SURCOUCHE/src with same name
